// src/components/Text/enums.ts
enum TextVariant {
  TitleMainWhite = 'TitleMainWhite',
  TitleMainPink = 'TitleMainPink',
  Heading1White = 'Heading1White',
  Heading1Grey = 'Heading1Grey',
  Heading1GreyDarkRegular = 'Heading1GreyDarkRegular',
  Heading1BlueRegular = 'Heading1BlueRegular',
  Heading2White = 'Heading2White',
  Heading2GreyRegular = 'Heading2GreyRegular',
  Heading2GreyBold = 'Heading2GreyBold',
  Heading2GreyDarkBold = 'Heading2GreyDarkBold',
  Heading2RedBold = 'Heading2RedBold',
  Heading4GreyDarkRegular = 'Heading4GreyDarkRegular',
  Heading4White = 'Heading4White',
  Heading4Pink = 'Heading4Pink',
  Heading4PinkBold = 'Heading4PinkBold',
  Heading4GreyMediumRegular = 'Heading4GreyMediumRegular',
  Heading4GreyMediumBold = 'Heading4GreyMediumBold',
  Heading4PinkBold = 'Heading4PinkBold',
  Heading4BlueBold = 'Heading4BlueBold',
  Heading4WhiteBold = 'Heading4WhiteBold',
  Heading4GreyBold = 'Heading4GreyBold',
  Heading4GreyDarkBold = 'Heading4GreyDarkBold'
}

enum TextFontFamily {
  Regular = 'proximaNova',
  Bold = 'proximaNova-Bold',
  Medium = 'proximaNova-Medium',
  Light = 'proximaNova-Light',
  SemiBold='proximaNova-SemiBold'
}

enum UnderlinePosition {
  Left = 'left',
  Center = 'center',
  Right = 'right'
}

export { TextVariant, TextFontFamily, UnderlinePosition };